Reduce the width of the infobar
When nagios or merlin isn't running and the 
database hasn't been updated in a while, the
info message at the top of ninja would cover
both logout and the search field.

This patch makes the error message somewhat 
smaller making it possible to still access
those fields.

This fixes issue #4136

Signed-off-by: Per Asberg <perasb@op5.se>