Handle possibly new default values for widgets
The database schema for oracle still sets the 
default values for username and setting to ' ' 
(a space instead of an empty string). 
The more correct solution would probably
be to modify the default values to NULL in all
database schemas but until then, this solution
will suffice (checkin both '' and ' ').

Signed-off-by: Per Asberg <perasb@op5.se>