Move htpasswd_importer into ninja
Previously we had to fetch users from htpasswd 
prior to logging on and therefore we needed a 
script that could be run from CLI. This is no
longer the case as users are synced upon a login
attempt.

This has the benefit of us not having to create
more database access methods than necessary.

Signed-off-by: Per Asberg <perasb@op5.se>