models/downtime: Fix crash in cron scripts
When running the cron scripts for scheduled downtime, they will always
run as the first user they could find, which might not have access to
all hosts that should be scheduled. This isn't a problem when you have a
monitor user, as that will be the first user, and that user will have
access to anything, but if you don't, things sometimes breaks.

This is worked around by ignoring any per-user restrictions when
scheduling downtime from the CLI.

This doesn't actually fix bug #4218, but at least prevents crashes.

This patch also makes check_if_scheduled return true for
{host,service}groups that are already scheduled - it used to always
return false (at least with a strict enough PHP).

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>