ninja_widget: Make widget settings work with mysql backend again
This is a evil hack that should be removed post release.

MySQL/PDO's support for quoting is buggy, so we don't want to rely on
that as we move forward - therefore we're using prepared statements
instead in the widgets. However, as our MySQL/PDO backend is
less-than-perfect an we have a release comming up, we've reverted to the
plain MySQL backend for this release, and the plain MySQL driver doesn't
support prepared statements at all.

So, for a limited time only, we have two queries for the price of one -
whenever the prepared statement fails, fall back to proper string quoting.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>