url helper: Don't crash when relative url is crazy
This fixes bug #4234, where a service description contained a : at a bad
place, which was used in a html::anchor call, which tried to decompose
the URL to reconstruct it again, which didn't work since it freaks
parse_url out.

The workaround constructs a valid, complete URL when parse_url fails, so
that it can then let nacoma deconstruct and reconstruct it without
failing.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>