reports model: Include each host once in uptime_queries for servies
Previously we added all services' hosts no matter if the host was
already added to the "select host downtime also for services" part
of the query. With this patch, we include each host exactly once,
which is ofcourse a lot better and protects against bugs and weird
stuff in various db servers.

Signed-off-by: Andreas Ericsson <ae@op5.se>