reports: Use timestamps in links to member's reports
This fixes bug #4244.

By not converting the timestamp to a date, we don't crash when
custom timeperiods are shorter than 24 hours.

Note that there can be a difference between the incoming timestamp and
the one in the link of up to 60, though, as the seconds are thrown away
before the report is generated.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>