db: Make pdomysql the default db driver (again)
I can no longer reproduce the errors discovered during testing for ninja
1.1.6 that caused us to revert to the mysql driver for the release.

Now that it is out, we should again try to use the pdomysql driver,
and then fix any problems that we find.

This patch also implements support for using the old/standard database
configuration parameters for configuring pdomysql instead of the ugly
dsn parameter.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>