Don't assign role to already existing user_id
If a user has been deleted from the nagios config
and another user has been added, we might end up
trying to assign a login role for this user that
already exists. This patch checks if this is the 
case before trying to insert into roles_users to
prevent the db from complaining.

Signed-off-by: Per Asberg <perasb@op5.se>