Remove the use of GET for multi commands
This patch removes the use of GET parameters
when requesting a multi object command.
Instead of passing the objects to multi_action()
and then calling submit() via the browser, all
objects are passed directly to the controller
method.

This has the benefit that all info is passed 
as POST which solves issues #4278 and #4219 

Signed-off-by: Per Asberg <perasb@op5.se>