Don't add extinfo.js if already added
When comments are loaded the javascript resource 
was added twice resulting in some toggle()
calls to take each other out resulting in
some elements remaining hidden.

This affected the show/hide of contactgroup
members on the extinfo details page for one.

This patch checks that the js file isn't already
added before trying to add it to the master
template.

Signed-off-by: Per Asberg <perasb@op5.se>