i18n: Handle upgrading old ninjas when serialized object isn't ascii
When upgrading from an older ninja that defaulted to communicating with
mysql as UTF-8 to a newer that expects to find UTF-8 encoded data in
latin1 ("bytestream") tables, any serialized object containing non-ascii
characters would cause a backtrace. To prevent this, try to manually
decode the string from utf8 one extra time to get the byte counts to add
upp. Also try to encode the string into utf8 in case someone does the
opposite.

This fixes bug #4242.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>