Display progress indicator for send and delete
When clicking on either send or delete for a 
scheduled report, nothing used to happen
until the jgrowl message popped up. This probably
resulted in users clicking multiple times or at
least wondering if anything was happening.
This patch fixes this by replacing the icons
temporarily with an animated rotating gif to
indicate that something actually is happening.

Signed-off-by: Per Asberg <perasb@op5.se>