comment model: Die, scary code, die!
fetch_comments was not only written in an quite convoluted way, but it
was also completely buggy: it failed to show users the comments to a
service if they lacked privileges to the master host, loaded all rows
from some tables it never actually used for anything, and was generally
unpleasant.

Rewrite to be less messy/to work.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>