Match local node with instance_id instead of name
ocimp and import.php provides different names for the local node,
but the instance_id of "self" has always been 0. Matching against
instance_id = 0 when looking for status on the local node makes a
lot more sense in light of that. It's also slightly faster, even
though that hardly matters in a table with hardly ever more than
10 entries.

Signed-off-by: Andreas Ericsson <ae@op5.se>