Let the last 2 widgets reuse model instance
This patch makes the last 2 widgets to reuse the
current_status model that is passed as an argument
from the tac controller. This will save us quite 
a lot of unnecessary queries and together with the
optimized outages model, this will speed up tac
quite a bit on larger systems. 

Signed-off-by: Per Asberg <perasb@op5.se>