nagios_auth: Make sure $this->id is always a number
If a user would be logged in, without being a contact, $this->id would
be false, and false would be cast to an empty string, which
could cause errors in random SQL queries.

Especially, this would cause errors in get_host_status in the host model
when a user was allowed to log in due to being
authorized_for_all_services, but without being a contact.

No contact should have id -1, so the result should be an empty result set
instead of a Kohana stack trace.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>