Implement search and display for notes
This patch implements the possibility to search
for and to display host- and service notes on 
status pages and search page.

It will not, however, introduce a new search flag
to only search for notes.

The My Account page is also updated to make it 
possible to select if the notes should be displayed
or not and how much of the notes that will be shown
in the listings. The entire note is always possible
to display by clicking or hovering it.

This implements issue #4146 

Signed-off-by: Per Asberg <perasb@op5.se>