models/reports: Only consider soft/hard for the primary event type
Downtime and start and stop events always log their entries with a
"hard" value of 0, meaning they were never considered when we're
asked to ignore soft states.

This patch moves the condition so that the 'hard' column is only
ever considered for object entries and all other entries are always
displayed.

Signed-off-by: Andreas Ericsson <ae@op5.se>