configuration: Use nacoma hook to make it possible to link/bookmark
As all nacoma subpages had the same URL, it used to be impossible
to refer to them specifically. Now, show nacoma's URL in the ninja URL,
so it's possible to bookmark, copy-paste, etc. While doing so, we also
must add the possibility to ninja to open an arbitrary nacoma page with
arbitrary parameters.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>