models/notifications: Rewrite query for limited contacts
The old query used unions over the notification table, which can be
pretty big and thus this operation can be slow. If the user is
authorized for loads of hosts and/or services, then mysql would also
need to create two temporary tables for all of those. The end result is
a bloody slow query.

This rewrite should be way faster, I hope.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>