Added Ninja automated testsuite
Added scripts to be able to run automated command-
line tests on ninja. They do, however, require that
ninja is properly installed and that we have a
working database connection etc.

The tests should be invoked by running
sh /path/to/ninja/test/ci/testsuite.sh 
or
sh /path/to/ninja/test/ci/testsuite.sh 
	/path/to/ninja 
	/path/to/ninja/test/ci/limited_tests.txt

The actual tests to run is defined in ninjatests.txt
and limited_tests.txt and adding more tests should
be quite straight forward.

These scripts are still in need of quite some 
optimization but it will have to do for a start.

Signed-off-by: Per Asberg <perasb@op5.se>