kohana db driver: When escaping integers, put quotes around them
This would have been a completely moronic thing to do, if PHP wasn't
such a moronic language.

The statement array("123"=>"123") is converted to array(123=>"123"),
that is, a string representation of a number in the key is automatically
cast to an integer. If we're using a proper database that doesn't
auto-quote for us when it gets bored (such as Oracle), it will not like
to see unquoted strings, resulting in potential crashes every time a
user-provided string is used as an array key.  Neither Oracle nor Mysql
mind getting numbers quoted (but I'm sure there is a database out there
that does care, that is going to go crazy about this, and that I'll have
to add support for at some point).

As PHP likes to do this "helpful" thing whenever it can, we have no way
of knowing if a number was supposed to be a string, so always quote.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>
(cherry picked from commit f3f43efbb9e4a29f18042865d33eb164200eb6bb)