widgets/tac_problems: Always filter on {SERVICE,HOST}_CHECKS_ENABLED
The old conditional was broken, as it didn't check on page '*' so it
always returned false.

Fixing it would also be broken, as that would include disabled checks,
which isn't included in the count from Current_status_Model, and would
thus be inconsistent.

As of last commit, the filters mentioned above does The Right Thing™ by
including passive checks when the user has chosen to, and not including
them otherwise, so always having it is suddenly the correct thing to
do.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>