models/notifications: Make object counting somewhat less insane
Retrieving most columns from *every* notification *ever*, and then
counting the number of rows retrieved, is not exactly a smart way to get
the number of rows. This was done every time the notifications page were
shown, merely to get the number of pages for the pagination to look
right.

*cringe*

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>