notifications: Once again, try to speed up object counting
It seems the last change was not enough.

This patch has the unfortunate side effect of not showing service
notifications for limited users when they are contact for a host but not
for it's services. I feel the need to do this, as not doing so has a
huge performance impact.

This is really a bug in import.php, one that most of ninja is affected
by already, and one that is fixed in ocimp already.

This is pushed to maint, as the slow object counting can make mysql
freeze for long periods of time, making merlin unable to push status
updates and rendering ninja unusable.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>