Remember sort options in session
This patch makes ninja to remember the users sort
order throughout a session. The sort order will be 
remembered for each status/host or status/service, 
that is, controller/method. It will keep sorting 
in session for anything that isn't a GET param as 
key. (status/host/<HOST_NAME> but not
status/host?host_name=<HOST_NAME>).

This implements issue #4260

Signed-off-by: Per Asberg <perasb@op5.se>