Added external widget support
This patch contains controllers, config and views
required to set up and use a ninja widget on an 
external web.

The user defined in config/external_widget.php
will only be able to view the external widget
which means it has to be a dedicated user.

Example iframe code is added to the comment in the
config file.

This implements issue #4589

Signed-off-by: Per Asberg <perasb@op5.se>