controllers/status: Added some method wrappers
Some of the links in the menu has quite a lot of
parameters which makes it quite hard for users to 
figure out how to access them and the menu
unnecessarily complex. 

This patch contains method wrappers for unhandled,
host, and service problems. The menu could hence
be updated to make highlighting in the menu easier
etc.

Signed-off-by: Per Asberg <perasb@op5.se>