Added new minimal master template: noc
This introduces a new minimal master template
intended to be used on NOC screens and it has
very small overhead when it comes to menu and
such. The menu is located in the vertical top
bar.

It also introduces a new tac variant which means
that it is now possible to have 2 separate ninja
windows open, each displaying different widgets.
This page has it's own settings on which widgets 
to show and the settings for each of them.

It is still to be regarded as experimental though.

Signed-off-by: Per Asberg <perasb@op5.se>