Pass current URI to current_time()
The current_time() method in the ajax controller
will now get the URI that the user was trying to 
load when the session expired. This way we will at 
least get one chance to reload the page that the 
user was thrown away from.

To make this more robust, we should add a form 
element to the login form that can hold the 
requested URI or the user will be lost anyway
once the current session has expired again.

Signed-off-by: Per Asberg <perasb@op5.se>