LDAP auth driver: Add LDAP escaping
This is sort-of semi-related to bug #4590 where crashes were caused by
unescaped input.

This is also a security fix, as it prevents LDAP injection. We're
already reasonably secure, as ninja doesn't talk to LDAP unless it has a
user with that username already, and ninja isn't really designed for
generic, public registration, but it surely won't hurt.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>