Prepare tac_problems widget for multiple instances
This patch may serve as an example on how widgets
supporting multiple instances could be done.

One of the important things here is that HTML IDs 
should be made unique by using the concatenated 
widget name and instance id. Most of the other
things needed is handled by the widget helper 
class when a widget is added.

Partial fix for issue #4488

Signed-off-by: Per Asberg <perasb@op5.se>