Verify widgets in widget_order
We have to make sure that we always have all user 
widgets for a page in the widget_order string or 
we might be unable to load a widget and get jGrowl 
messages about this. 

This also makes the add_widget() to also add
the widget in the widget_order string which will
help users uploading new widgets.

Signed-off-by: Per Asberg <perasb@op5.se>