command: Handle _host-too in group cases
The list of events that can have _host-too was taken from the nagios
cgi's, so it's actually a superset of the list Ninja allow you to set
this for. Should be good to have, anyway.

This fixes bug #4529.

This commits also fixes the wording issue raised in the bug, as well as
a pluralization issue that caused _host-too not to work for
*_HOST_SVC_CHECKS.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>