alert history: Use database instead of log files
This is a fix for bug #4568. It also fixes bug #3231 (well, half of it)
and, I guess, bug #4365 (though I guess "clean" is in the eye of the
beholder, and I don't really trust my eyes).

It doesn't implement all the General options that the old alert log had,
as not all events are in the database. In particular, I'm thinking
flapping alerts probably has as much reason to be in an alert log as
downtime alerts, though it's probably for the best to get rid of
logrotation, event handlers, and external commands. Still, perhaps the
merlin log importer has to be modified. Oh, and we only get the first
soft alert per object - I think we decided that was a feature?

Some of the roughness is due to bug #4566, free text filters. We should
implement that one with some kind of ninja-wide grid system, so I'd
prefer to wait until after the vacations are over, and then talk about
the best way to do that. Thus, the view will have to be rewritten then,
so I'm not gonna bother prettying it up too much.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>