templates: Stop doing that weird capitalization thing
Title case is a thing, but as the name implies, it's for titles, not
everything that's in an a tag. Both bug #4491 and #4554 contains
complaints about this.

We've also been doing it completely inconsistently, so if I didn't find
all cases, well, it wasn't that much better to begin with.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>