Add Nacoma hook, handle renames of objects in scheduled reports
The hook detects renames and deletions to objects from nacoma, and calls
into ninja for each object it finds to let it clean up.

The rename handler so far handles changing object names in saved
reports. The deletion handler is empty.

There are no tests here for this (yet), but some will be added to the
autotest repo.

This fixes bug #4554.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>