auth methods: Be consistent about auth.driver/auth.auth_methods
auth.driver is *only* to be trusted with telling you the authentication
method used for this particular session - nothing else.

auth.auth_methods lists all available authentication methods. This can
now be a simple string value, like the old auth.driver setting, or an
array.

Never set auth.driver in config/auth.php. It is still supported for
existing users, but is as of now deprecated. It's supposed to be set
automatically by the ninja constructor.

Take this state of affairs into account in cli/insert_user_data to make
sure all users are always imported into the database.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>