comment/downtime model: Don't check host authorization for services
The reason for this change is that limited users sees all comments twice
before this patch.

If a user is authorized for the host, (s)he is authorized for the
service. If a user is authorized for the service, contact_access is
supposed to say so.

I think there are/were bugs in import.php that makes it not always so,
but I think the authorization queries for the service itself breaks when
that is not the case anyway, so let's remove this.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>