Kohana error page: Write error stacktrace to file instead of user
Instead, only the path to the error information is printed. This has
several advantages:
 * It should make it easier for the local op5 admin to send the full
   debug information, instead of just something like a screenshot.
 * It hopefully looks less scary.
 * It will never show the user database passwords or other confidential
   information.

The disadvantage is that the admin must now download the information
from the server - it would be pretty cool if they could view it through
the web, but that creates a new can of worms. Later!

To aid development, this will be disabled on development versions, where
everything will be as before.

I now consider bug #4518 completely closed.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>