Fixed trends graph
The trends graph didn't show the small slices 
correctly in some browsers when getting < 1% of
the total width. Fixed this by setting those slices 
to 1px instead. The graph will never be totally
accurate when it comes to the bottom row (time 
information) and the slices especially when showing 
a large number of small events. This will have to 
be fixed by switching to a server side created
graph that will have to wait for now.

This fixes issue #4006

Signed-off-by: Per Asberg <perasb@op5.se>