Added 403 + 404 error code handling
We used to handle the case when a user entered a
bad method (page) for an existing controller
but not when trying to access a non existing
controller. This patch fixes this by adding
new 403/404 messages. Could probably be enhanced
quite a bit but they will do for now.

Partial fix for issue #3970

Signed-off-by: Per Asberg <perasb@op5.se>