reports: Fix handling of nested downtimes
When a host or service ends up in two downtime periods at once, merlin
(but not import) will shove both of those downtime start and downtime
end events in the database. This situation should not break our reports.

We fix this by completely ignoring the downtime_depth in the table, as
it only contains lies and deceptions.

This commit does still not care about reports that start when we're in
two nested downtimes - that situation will be endlessly painful to fix,
and it literally is a special case of a special case, so I'll just
ignore it for now.

Add new report test for this, and also add new way of creating tests:
because this only happens in merlin, not when running the log import, it
is not possible to generate a log file to simulate this. We must thus
start with a SQL file instead.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>