reports: The big downtime shuffle
First, all downtime events are stored with state 0 (as in
"inapplicable"), which the cstate map handled correctly, but not the
log, so even though the breakdown was correct, the trend graph wasn't.
This meant our tests always said everything worked, when in fact it
didn't. If somebody spotted this, they might be led to look at the log
shown in some avail reports with these values, and note that it agreed
with the graph, giving them the wrong idea. Balls.

This is an actual fix to bug #3895 and bug #4238.

Secondly, the service breakdown on host availability report pages
ignored any scheduled host downtime, even though the service
availability reports took them into account, creating potentially
massive discrepancies. Handle this.

Thirdly, move the "handle scheduled downtime as uptime" code out further
away from the core code. It used to be handled in st_parse_line (where
it would replace the "real" state, which as described above was always
0, with 0. Oh, and it was handled in plenty of other places as well, for
good measure), which made it impossible to, when converting the state
table, figure out which lines this affected. Move this out, to st_log
appending and state table conversion, to make it possible to calculate
this value.

Fourthly, while we're at it, show this value in single host or single
service avail reports. This is a precursor and a start for implementing
bug #2257 - implementing it completely will be left for later.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>