reports: Hide events that are outside of report period from log/trends
This is actually somewhat trickier than you'd think.

We only hide events that are completely outside of our report period,
because if, say, a machine crashes before the report period starts, it
will still affect the availability numbers for the host, so we need to
include the part of the down event that happens during the report
period.

We could pretend that it happened right at the report start, but that
would suck, as the report reader might not associate the event in the
report with the real-world event that caused the problem if the
timestamps vary too much (on a "workhour" report it could turn a friday
night crash into a monday morning crash). For the same reason, we show
the full duration of events, even if they end after the timeperiod ends.

This means that the trends graph will only occasionally show holes, and
most of the log will be back-to-back events, no matter if the timeperiod
ends or begins, and availability numbers will still be calculated based
on only part of the shown time ranges. This might be surprising at
first, but I think it makes the most sense as soon as you grok it.
Future work might be to fade the parts of the trend graph that falls
outside of the timeperiod, but that is currently highly non-trivial to
do.

This fixes bug #4433.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>