reports: Re-fix downtime handling for services
If we create a report of many services, and their hosts go in and out of
scheduled downtime, the downtime events haven't been associated with the
services. The other day, I fixed so they would be, if we made a report
over services on one single host. This is only one of a lot of cases,
and this is supposed to fix most of the other.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>