reports: Handle timeperiods with excludes and exceptions
This adds supports for excluding an irregular timeperiod, but it also
fixes the support for a regular excluded timeperiod, when the "parent"
timeperiod has exceptions.

Somewhat hacky, but we'd need to rewrite the timeperiod model to make
this prettier, and I have no time for this right now.

This fixes bug #4297.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>