commands: Make it possible to configure the default values for commands
Some of them are configurable from within Ninja, some more are
configurable from a config file, and some aren't configurable at all, at
least not yet.

The values that are configurable from within Ninja are hopefully the
most common ones, the ones that can be configured in a config file are
probably the less common ones, and the ones that aren't configurable are
hopefully the impossible ones. We can relatively easily move options
between these groups in the future if my guesses are wrong.

This resolves bug #4322.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>