command: Add "services too" option for acknowledgements
As there is no "acknowledge all service problems on host" command, we
need to manually find all services, and ack them one by one. It's not
enough to only do this ninja-side, as we want notifications from nagios
to respect this as well.

I intentionally don't care about the service state: it seems reasonable
to not make sure the service is in the same state as the host, since a
host that is DOWN (or could quite reasonably result in services that are
both CRITICAL or UNKNOWN, and if the service for some reason is still
OK, nagios just ignores the ack.

This fixes bug #4027.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>