upload: Fix upgrades
Add a bit of extra fortifications for removals, but more importantly,
don't print a fail message when widget wasn't added on upgrade. Unless
something was terribly wrong before, the database will already have the
widget, so the add_widget call should return false. We still do the
call, though, as that should be safer in case of extraordinary
circumstances.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>