authenticated: Don't rely on javascript to enforce authentication
For obvious reasons.

Luckily, ninja simply crashed if you tried to access pages you were
unauthorized to see, so there was no information leaks, but still.

The reason the javascript was added in d56570a9 was a good one, though,
so put the equivalent in the login template itself: if the template
discovers that it's been loaded in an iframe, make it reload itself in
the whole window.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>