comment/scheduled_downtime: Rewrite multi-delete
This is a bit invasive at this point in the development cycle, but it
really was necessary, as the old code didn't actually work. It does the
following:

The "Select All" column was sortable. This fixes bug #4755.

The "Select All" column was impossible to check, because clicking on it
only triggered a sort. This fixes bug #4753.

The "Delete Selected" button was part of the table, and therefore was
sorted with all the other columns. This was mentioned in bug #4755 as
well.

The "Delete Selected" button was part of the table, and therefore
subject to being filtered away. I believe this fixes bug #4754.

The filter didn't properly deselect filtered buttons, making it possible
to delete comments/downtimes that were out of view.

The "Delete services too" button on host downtime didn't actually do
anything the regular delete button didn't.

The "Delete services too" button didn't verify that anything was
actually checked.

Make the view when there are no downtimes roughly the same for the host
part and the service part.

About 100 lines of js for multi-select functionality was copy-pasted
(down to the comment level) from the standard, tested, implementation
used everywhere else. Kill the duplicated code.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>