Restrict table fix for multi search to search result page
The tables gets broken when using a hidden td as
first element in table and showing on click like 
when using multiple select. This is solved by a
JQuery script but the problem with my last commit
was that it accidentally affected all pages, 
resulting in a thin grey vertical line appearing
in almost all tables. This patch amends by only 
applying the script to the search result table for 
comments.

Signed-off-by: Per Asberg <perasb@op5.se>