ninja_widget: Fix widget hidden detection
Widgets should only be hidden when the user has custom settings for it,
and those settings include ['status' => 'hide']. Before this patch, they
were hidden when the users settings (excluding hide settings) were
non-empty and didn't include this widget.

This patch ignores the case where the global default-setting is
['status' => 'hide'], but there's no way to do that without manual
database fiddling, and if you do that, you're on your own.

I hypothesize that the reason we didn't find this earlier is that we
at some point (I think) used to copy every single widget setting to the
user, we have test machines that were installed at that point, and our
other test machines haven't been tested properly with any other
settings than hide (and reordering).

This fixes bug #4809.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>