iframe-adjust: No, we don't want inner scroll bars
I don't like the sollution in cf22072a1209612d4e84d3f4491c14a41040ac2e
at all, as different iframes have way different views of what is "too
small", so rework. The real issue is that #content was too high, so
instead set #content to the same hight as it's content. This fixes the
issue with only one set of scroll bars.

Also put hypermap back - it was lost in a merge.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>