ninja_widget: Make it even harder to get "Widget not found" errors
Apparently, it is, due to weird upgrade patterns and randomness,
possible to end up with different set of "global" widgets from "local"
ones. If the two sets still shared the same length, this would result in
"widget not found" errors.

While I don't know exactly how it happened, this was found to be a
problem on beta64.

Fix by removing optimization to make all "global" widgets always try to
add themselves to the "local" widget set. We could rewrite this to be
smarter and faster by not treating these two arrays as different things,
but I'm in a hurry.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>