upgrade: Handle when scheduled_reports_db_version contains non-numeric
It contains the string "1.0.0" by default, which is not a number. As we
don't care by the minor or micro version, we'll just throw them away.

All other version tables have integer version columns, so it's no
problem there. Should scheduled_reports_db_version contain anything that
is not a number or a "version string", it will still break, but it
shouldn't do that.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>