templates: Use HTML5 doctype, not XHTML1 strict
Calling what we've got XHTML is just terribly, terribly broken: it
really must be parsed through a HTML parser, not an XML one, so use a
easy-to-remember doctype that means that.

Not that we haven't always used a HTML parser, but now the doctype
doesn't lie, so validators can actually tell what it is we want them to
validate, so I can use them to find errors. Convenient.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>