Makefile: Shush when generating test coverage
We were being loud and obnoxious, and nagios doesn't like that. Maybe,
if we try to behave and be polite, nagios will introduce us to it's
friends, and maybe then his friend PNP will be there, and perhaps
he'll like us and ask us to the prom and then we'll get married, have a
house, a dog, and lots of beautiful little graphs.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>