widgets: Add new error widget
There's currently this horrible string in the widget helper that
is being shipped off to the user when there's an error. Let's create a
real widget instead!

As it's not installed into the database, we don't have to worry about it
showing up at inconvenient times.

This is a newstyle widget. The newstyle API doesn't exist yet, but I
predict it will soon magically spring into existence.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>