Implement new widget API in Ninja
This breaks absolutely every widget. I've already commited porting
instructions, and will go do some actual porting myself ASAP.

One of the goals of the new API has been to reduce the amount of
cargo-cult coding needed to create a widget: the view is now completely
void of any magic at all, the controller has way less of it (unless
there's a gazillion options - not happy with that part of the API yet),
and as we add the chrome stuff at a central place, from the widget
system, rather than once from every widget, we should be able to make
bigger changes to the widget system in the future without having to
upgrade all the widgets again, and also increase consistency and
flexibility with regards to the look and feel. We could still be more
helpful with custom options, though.

Another goal was to make the model smarter, so we won't have to care so
much about things like user-specific vs global widgets and settings -
see bug #4809, which is solved here - as well as making widgets less
coupled to the page they're on, so any newly installed widget for the
tac will instantly be available on the noc as well - this used to not
work (while we're at it, throw away the noc controller noc view template
[as in, not the global noc view template, the controller specific one],
as it is an exact duplicate of the tac view template).

The reason for doing this in the first place was to enable multiple
instances of one widget, which is now also in place. The UI for copying
widgets - especially the copy widget icon - is super-temporary: I've
mostly been concerned with the backend.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>