view/command/request: Fix default array handling
We have a convenient array of possible values. When the 'default' array
is set, we don't use the possible values, but instead use only the
defaults. Whether this makes sense or not, I will not get into right
now.

What I will get into is that when rendering the 'default' array, we
would only use the array keys as both values and option text in the
select. We should of course use the label text from the value array.

This will make downtime multiselects not look like crap, as soon as I've
fixed them.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>