Remove anoying error message from multi_command:
If you forgot to change the command you wanted to send for
your host(s)/service(s)/hostgroup(s)/servicegroup(s) so
you sent the default "Select action" you got a error message
from the command controller.

We now supress this but not allowing you to send the command
to the controller if you forgot to change.

Signed-off-by: Mattias Ryrlén <mattiasr@op5.com>