Make more strict, rely on less PHP bugs
One of my favourite bugs in PHP - frankly, among many - is
https://bugs.php.net/bug.php?id=30355

For ninja, all helpers are called statically, and they sometimes access
$this, and as long as there's no strict declaration, $this will point to
the controller that called it (assuming it was actually called from a
controller, which is not a given).

We tend to use this to access the translate and session members as well
as the add_view method in the ninja controller from the helpers, even
though they're not really defined anywhere and relies on magic call
paths and luck to work ("luck", due to nothing of the sort being
documented anywhere).

This keeps causing confusion, as I'm not the only one who has foolishly
believed that I could modify the class I was in, and have the class
object reference in my methods reflect my modifications.  In fact,
several helpers contains completely useless constructors that are never
executed.

To try to get away from this, I've started running my dev machine with
error reporting "strict", as it tells you to add static if you use
something statically. This is a first batch of changes that makes it
possible to at least click around somewhat in that mode.

There's a new ninja helper that will let you access the add_view method.
As for the other variables, they're always stored in the zend registry,
that we only use to keep state and variables around globally, so we
can just, I dunno, use it.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>