ninja_unit_test: Add exit(0) to the end
There's probably better ways to do this, if one could penetrate the way
kohana's mind works, but for lack of jedi mind tricks, I'll just do it
the easy way. Without an exit(), ninja prints a 302 message after
testing is complete, which is quite unhelpful when we want to find the
output from the check.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>